pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["jerimkaura <jerimkaura001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-environ = "^0.8.1"
black = "^22.1.0"
flake8 = "^4.0.1"
pre-commit = "^2.17.0"
python-dotenv = "^0.20.0"
python-decouple = "^3.6"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
